authorization: setup everywhere

+++ src/locations/controller.ts
@@ -2,6 +2,7 @@ import { Router, Response, Request } from "express";
 import Service from "./service";
 import { Location, PartialLocation } from "./model";
 import { NotFoundError } from "../common/http_errors";
+import authorize from "../common/middlewares/authorize_middleware";
 
 const controller = Router();
 
@@ -25,39 +26,51 @@ controller.get("/:id", (req: Request, res: Response, next: Function) => {
     .catch((err) => next(err));
 });
 
-controller.post("/", (req: Request, res: Response, next: Function) => {
-  Service.createOne(req.body as Location)
-    .then((data: Location) => {
-      res.status(201).json(data);
-    })
-    .catch((err) => next(err));
-});
+controller.post(
+  "/",
+  authorize(["staff"]),
+  (req: Request, res: Response, next: Function) => {
+    Service.createOne(req.body as Location)
+      .then((data: Location) => {
+        res.status(201).json(data);
+      })
+      .catch((err) => next(err));
+  },
+);
 
-controller.delete("/:id", (req: Request, res: Response, next: Function) => {
-  Service.deleteOne(Number(req.params.id))
-    .then((id) => {
-      if (id === null) {
-        throw new NotFoundError(
-          `Could not find location with id ${req.params.id}`,
-        );
-      }
+controller.delete(
+  "/:id",
+  authorize(["staff", "owner"]),
+  (req: Request, res: Response, next: Function) => {
+    Service.deleteOne(Number(req.params.id))
+      .then((id) => {
+        if (id === null) {
+          throw new NotFoundError(
+            `Could not find location with id ${req.params.id}`,
+          );
+        }
 
-      res.status(204).json();
-    })
-    .catch((err) => next(err));
-});
+        res.status(204).json();
+      })
+      .catch((err) => next(err));
+  },
+);
 
-controller.patch("/:id", (req: Request, res: Response, next: Function) => {
-  Service.updateOne(Number(req.params.id), req.body as PartialLocation)
-    .then((data: Location | null) => {
-      if (data === null) {
-        throw new NotFoundError(
-          `Could not find location with id ${req.params.id}`,
-        );
-      }
-      res.status(200).json(data);
-    })
-    .catch((err) => next(err));
-});
+controller.patch(
+  "/:id",
+  authorize(["staff", "owner"]),
+  (req: Request, res: Response, next: Function) => {
+    Service.updateOne(Number(req.params.id), req.body as PartialLocation)
+      .then((data: Location | null) => {
+        if (data === null) {
+          throw new NotFoundError(
+            `Could not find location with id ${req.params.id}`,
+          );
+        }
+        res.status(200).json(data);
+      })
+      .catch((err) => next(err));
+  },
+);
 
 export default controller;
