+++ src/reservations/controller.ts
@@ -2,28 +2,37 @@ import { Router, Response, Request } from "express";
 import Service from "./service";
 import { Reservation, PartialReservation } from "./model";
 import { NotFoundError } from "../common/http_errors";
+import authorize from "../common/middlewares/authorize_middleware";
 
 const controller = Router();
 
-controller.get("/", (_req: Request, res: Response, next: Function) => {
-  Service.getAll()
-    .then((data: Array<Reservation>) => res.json(data))
-    .catch((err: Error) => next(err));
-});
+controller.get(
+  "/",
+  authorize(["staff", "customer", "owner"]),
+  (_req: Request, res: Response, next: Function) => {
+    Service.getAll()
+      .then((data: Array<Reservation>) => res.json(data))
+      .catch((err: Error) => next(err));
+  },
+);
 
-controller.get("/:id", (req: Request, res: Response, next: Function) => {
-  Service.getOne(Number(req.params.id))
-    .then((data: Reservation | null) => {
-      if (data === null) {
-        throw new NotFoundError(
-          `Could not find reservation with id ${req.params.id}`,
-        );
-      }
+controller.get(
+  "/:id",
+  authorize(["staff", "customer", "owner"]),
+  (req: Request, res: Response, next: Function) => {
+    Service.getOne(Number(req.params.id))
+      .then((data: Reservation | null) => {
+        if (data === null) {
+          throw new NotFoundError(
+            `Could not find reservation with id ${req.params.id}`,
+          );
+        }
 
-      res.json(data);
-    })
-    .catch((err) => next(err));
-});
+        res.json(data);
+      })
+      .catch((err) => next(err));
+  },
+);
 
 controller.post("/", (req: Request, res: Response, next: Function) => {
   Service.createOne(req.body as Reservation)
@@ -33,31 +42,39 @@ controller.post("/", (req: Request, res: Response, next: Function) => {
     .catch((err) => next(err));
 });
 
-controller.delete("/:id", (req: Request, res: Response, next: Function) => {
-  Service.deleteOne(Number(req.params.id))
-    .then((id) => {
-      if (id === null) {
-        throw new NotFoundError(
-          `Could not find reservation with id ${req.params.id}`,
-        );
-      }
+controller.delete(
+  "/:id",
+  authorize(["staff", "owner", "customer"]),
+  (req: Request, res: Response, next: Function) => {
+    Service.deleteOne(Number(req.params.id))
+      .then((id) => {
+        if (id === null) {
+          throw new NotFoundError(
+            `Could not find reservation with id ${req.params.id}`,
+          );
+        }
 
-      res.status(204).json();
-    })
-    .catch((err) => next(err));
-});
+        res.status(204).json();
+      })
+      .catch((err) => next(err));
+  },
+);
 
-controller.patch("/:id", (req: Request, res: Response, next: Function) => {
-  Service.updateOne(Number(req.params.id), req.body as PartialReservation)
-    .then((data: Reservation | null) => {
-      if (data === null) {
-        throw new NotFoundError(
-          `Could not find reservation with id ${req.params.id}`,
-        );
-      }
-      res.status(200).json(data);
-    })
-    .catch((err) => next(err));
-});
+controller.patch(
+  "/:id",
+  authorize(["staff", "owner", "customer"]),
+  (req: Request, res: Response, next: Function) => {
+    Service.updateOne(Number(req.params.id), req.body as PartialReservation)
+      .then((data: Reservation | null) => {
+        if (data === null) {
+          throw new NotFoundError(
+            `Could not find reservation with id ${req.params.id}`,
+          );
+        }
+        res.status(200).json(data);
+      })
+      .catch((err) => next(err));
+  },
+);
 
 export default controller;
